Update README.md to include a new hardhat template
I created a hardhat template that helps allows for users to create custom Uniswap V4 Hooks. It includes functionality to mine addresses as well. I believe that this would be a useful addition to the list because there is a lack of examples of Uniswap V4 hooks in hardhat so this helps to fill that gap.